+++ packages/google_ml_kit/example/lib/activity_indicator/activity_indicator.dart
@@ -1,0 +1,77 @@
+import 'package:flutter/material.dart';
+
+class Toast {
+  void show(String message, Future<String> t, BuildContext context,
+      State<StatefulWidget> state) async {
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    showLoadingIndicator(context, message);
+    final verificationResult = await t;
+    Navigator.of(context).pop();
+    if (!state.mounted) return;
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+      content: Text('Result: ${verificationResult.toString()}'),
+    ));
+  }
+
+  void showLoadingIndicator(BuildContext context, String text) {
+    showDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (BuildContext context) {
+        return WillPopScope(
+            onWillPop: () async => false,
+            child: AlertDialog(
+              shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.all(Radius.circular(8.0))),
+              backgroundColor: Colors.black87,
+              content: LoadingIndicator(text: text),
+            ));
+      },
+    );
+  }
+}
+
+class LoadingIndicator extends StatelessWidget {
+  final String text;
+
+  const LoadingIndicator({Key? key, required this.text}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        padding: EdgeInsets.all(16),
+        color: Colors.black.withOpacity(0.8),
+        child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
+            children: [_getLoadingIndicator(), _getHeading(), _getText(text)]));
+  }
+
+  Widget _getLoadingIndicator() {
+    return Padding(
+        child: SizedBox(
+            child: CircularProgressIndicator(strokeWidth: 3),
+            width: 32,
+            height: 32),
+        padding: EdgeInsets.only(bottom: 16));
+  }
+
+  Widget _getHeading() {
+    return Padding(
+        child: Text(
+          'Please wait …',
+          style: TextStyle(color: Colors.white, fontSize: 16),
+          textAlign: TextAlign.center,
+        ),
+        padding: EdgeInsets.only(bottom: 4));
+  }
+
+  Widget _getText(String displayedText) {
+    return Text(
+      displayedText,
+      style: TextStyle(color: Colors.white, fontSize: 14),
+      textAlign: TextAlign.center,
+    );
+  }
+}